Give warning about false positives from case sensitive wikis
If on a case sensitive wiki (wiktionary) someone does [[File:foo.png]],
where Foo.png is a file on commons, foo.png will be recorded in
the globalimagelinks table, but Foo.png will be used. As a result,
foo.png will incorrectly show up as wanted.

See bug 71359 for full issue. This patch attempts to reduce severity
of the issue by adding a warning to the top of the page, and
rendering entries for foo.png as File:foo.png, instead of the
current behaviour of converting to the canonical File:Foo.png
form.

Change-Id: Iaf176c4a71fdca5b026470338fcd6aaee0926f2e