Fix database query from api for number-only titles
When requesting a title with only number (like File:94 from the task),
the number-like string is converted to an integer when used as array
key. WHERE gil_to = 94 is executed from the database and every row is
converted to number and then compared, which makes the query very slow
and where the result can contains other titles as requested and therefor
the pageids is not found in the array.
This also produce a continue tag in the api as there is more results to
found while actually the query should not find anything, because the
file is missing and the name is invalid (missing file extension)

Bug: T263601
Change-Id: I55fd83f5fbf9bc0e9b2496c99649138a081f211e